Finalized cleanup for backend
