+++ server/index.js
@@ -2,7 +2,7 @@ import express from 'express';
 import cors from 'cors';
 import mongoose from 'mongoose';
 import dotenv from 'dotenv';
-import routes from './api.js';
+import routes from './routes/api.js';
 
 // Load environment variables from .env file
 dotenv.config();
